requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pdfplumber==0.11.0
python-multipart==0.0.9
python-dotenv==1.0.1
openai==1.30.1
pydantic==2.7.1

# Optional: for DOCX resume support
# python-docx==1.1.2

# Optional: switch to PostgreSQL
# psycopg2-binary==2.9.9
